docs: add license
